finally its working while separating files and db connection
This is the first time I have done something like this, its a great feeling after all.

I didn't think that it would work without installing mongoose or another packages. But it finally did.

The trick is, copy the whole mongodb.connect() function into a single variable and export that variable. After that you can use it just  by requiring as if it was written above the code where you are using db to do CRUD operations.