ID validation of single blog page.
Still the remarks are left to describe in the code.

Its a great challenge to configure the validation of the single blog page, the "isValid" function of mongoDB still doesn't check the entered ID exists in Database or not, it just checks the length of the ID string.

But if someone enters a fake ID string with same length it counts it as real one, and then tries to load it.

In this commit, I've corrected these problems,

 1. Check first if the id is long enough to be an actual ID.

 2.  Check inside Database, if that entered long string is actually exists or not. If not, then it shows error message and if so, it will load the certain data in a page.